Ask for a token, not a password
According to [this comment][ref], the usage of a password is currently
broken but it is possible to pass an access token instead.

This access token is then stored in the configuration file, so adjust
the message to ask for an access token, tell the user where then can
generate a new one, provide the required scope, qnd do not tell that the
value is not stored.

[ref]: https://github.com/github/hub/issues/2655#issuecomment-735836048